+++ src/HighTable.tsx
@@ -2,22 +2,22 @@ import React, { MouseEvent, useCallback, useEffect, useMemo, useRef, useState } 
 import Cell from './components/Cell.js'
 import Row from './components/Row.js'
 import RowHeader from './components/RowHeader.js'
-import { DataFrame } from './dataframe.js'
+import { DataFrame } from './helpers/dataframe.js'
+import { PartialRow } from './helpers/row.js'
+import { Selection, SortIndex, areAllSelected, computeNewSelection, isSelected, toggleAll } from './helpers/selection.js'
+import { OrderBy, areEqualOrderBy } from './helpers/sort.js'
 import { rowError } from './helpers/text.js'
 import { useInputState } from './hooks/useInputState.js'
-import { PartialRow } from './row.js'
-import { Selection, SortIndex, areAllSelected, computeNewSelection, isSelected, toggleAll } from './selection.js'
-import { OrderBy, areEqualOrderBy } from './sort.js'
 import TableHeader, { cellStyle } from './TableHeader.js'
 import { stringify as stringifyDefault } from './utils/stringify.js'
 import { throttle } from './utils/throttle.js'
-export { arrayDataFrame, sortableDataFrame } from './dataframe.js'
-export type { DataFrame } from './dataframe.js'
-export { asyncRows, awaitRow, awaitRows, resolvableRow } from './row.js'
-export type { AsyncRow, Cells, PartialRow, ResolvableRow, Row } from './row.js'
-export { rowCache } from './rowCache.js'
-export type { Selection } from './selection.js'
-export type { OrderBy } from './sort.js'
+export { arrayDataFrame, sortableDataFrame } from './helpers/dataframe.js'
+export type { DataFrame } from './helpers/dataframe.js'
+export { asyncRows, awaitRow, awaitRows, resolvableRow } from './helpers/row.js'
+export type { AsyncRow, Cells, PartialRow, ResolvableRow, Row } from './helpers/row.js'
+export { rowCache } from './helpers/rowCache.js'
+export type { Selection } from './helpers/selection.js'
+export type { OrderBy } from './helpers/sort.js'
 export { resolvablePromise, wrapPromise } from './utils/promise.js'
 export type { ResolvablePromise } from './utils/promise.js'
 export { HighTable }
