+++ test/helpers/row.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
-import { AsyncRow, Row, awaitRows } from '../src/row.js'
-import { wrapPromise } from '../src/utils/promise.js'
+import { AsyncRow, Row, awaitRows } from '../../src/helpers/row.js'
+import { wrapPromise } from '../../src/utils/promise.js'
 
 export function wrapObject({ index, cells }: Row): AsyncRow {
   return {
